# packages/api-schema/openapi/core.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PERUM Control Plane",
    "description": "Manages organizations, provisioning, billing and observability for PERUM tenant stacks.",
    "version": "0.5.3"
  },
  "paths": {
...
    "/api/support/tickets": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "My Tickets",
        "operationId": "my_tickets_api_support_tickets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response My Tickets Api Support Tickets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "support"
        ],
        "summary": "Open Ticket",
        "operationId": "open_ticket_api_support_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Open Ticket Api Support Tickets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/support/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "Org Ticket",
        "operationId": "org_ticket_api_support_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Org Ticket Api Support Tickets  Ticket Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/support/tickets/{ticket_id}/messages": {
      "post": {
        "tags": [
          "support"
        ],
        "summary": "Org Reply",
        "operationId": "org_reply_api_support_tickets__ticket_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Org Reply Api Support Tickets  Ticket Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "MessageCreate"
      },
...
      "TicketCreate": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Subject"
          },
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "message"
        ],
        "title": "TicketCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}